simulations/1-full-engagement/calder-insurance/02_week2_solution/prototype/requirements.txt
anthropic>=0.40.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pytest>=8.0.0
